Add compile-time option to enable MSP push messages over the VCP port Enables easier development and testing of MSP displayPort devices as the existing VCP connection to a computer can be used.
To use, define `USE_MSP_PUSH_OVER_VCP` either in the code or when compiling like:
```
make TARGET=STM32F405 OPTIONS=USE_MSP_PUSH_OVER_VCP
```

The existing code has the VCP blocked for MSP push serial as it can cause a number of problems. If there is no consumer then the USB endpoint will quickly fill its buffer and start flow-controlling. This leads to the the STM libraries blocking and the firmware wedges until data is read. So it's only appropriate that this is used for development testing.