Sets antigravity gain to 0 when the user wants none.
In the PR adding P to antigravity, the default value was changed so that zero now meant no antigravity, whereas before 1000 meant none.  Unfortunately, `ITERM_ACCELERATOR_GAIN_OFF` remained at 1000.  This was an oversight.
Note also that the default value of 3500 now results in more antigravity effect than previously; it is equivalent to 4500 in the previous system.