Validate that the battery min cell voltage is less than the max cell voltage Resets to defaults if not.
If the min/max were set to the same value then this will lead to div-by-zero errors.

If the min was set to a greater value then this would lead to negative values and underflow in the battery remaining calculations. This would affect the OSD, dashboard, various telemetries, LED strip, and the Configurator.