RTFL: Remove PID gain normalisation
The effect of headspeed on PIDF gains is much more complex
than just a scaling effect by RPM^2. Also, there are limits
how much the PID gains can be changed, different on every
helicopter model.

The best option is to manually tune the PIDF gains for each
headspeed, in separate profiles.

If using only one profile, the PIDF gains should be tuned
for the worst case, which is the highest headspeed used.