Fix gyro 1k ODR handling
The slowest MCUs, F411 and G474 can't keep up with the 8k gyro rate,
so they are using a lower gyro ODR, 1k or 2k.

1k ODR is implemented differently for each gyro model.

MPU6000 and MPU6050 support ODR divider with 8k and DPLF=0 (250Hz LPF),
so they can use the divider.

MPU6500, MPU9250 and ICM20689 don't support the divider with 8k, so
they must use 1k ODR with DLPF=1 (184Hz LPF).

ICM42688P support multiple ODRs directly, including 1k, with multiple
LPF settings.

The gyro detection interrupt count is dropped to 100, as 1000 was
causing a timeout and gyro detection failure.

The gyro rate monitoring period is dropped from 25000 INTs to 10000,
because the former was causing a uint32 overflow with faster MCUs.

Co-authored-by: rtlopez <rtlopezdev@gmail.com>