Add gyro_rate_sync parameter
The gyro rate sync feature in the scheduler can be disabled
by setting this parameter to false.

Then the gyro task is run at a constant rate, *not* matching
the gyro rate. Essentially, it is doing a rate conversion from
the gyro (slightly varying) rate to the MCU gyro task rate.